Renaming vrrb_vrf and vrrbdb
