+++ crates/block/src/header.rs
@@ -10,7 +10,7 @@ use secp256k1::{
 use serde::{Deserialize, Serialize};
 use utils::{create_payload, hash_data};
 use versa_core::claim::Claim;
-use vrrb_vrf::{vrng::VRNG, vvrf::VVRF};
+use versa_vrf::{vrng::VRNG, vvrf::VVRF};
 
 use crate::{block::Block, InnerBlock, NextEpochAdjustment};
