+++ crates/cli/tests/whitelist_data/whitelist.json
@@ -1,1 +1,1 @@
-{"genesis-farmers":[{"kademlia_liveness_address":"127.0.0.1:13004","kademlia_peer_id":[185,28,187,248,214,245,246,193,253,220,97,7,202,198,244,9,67,64,212,39,230,117,23,84,160,245,14,95,204,154,162,54],"node_id":"node-4","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12004","udp_gossip_address":"127.0.0.1:11004","validator_public_key":"027f8995f19148a1fa4980e208e72ccab82769e670598378881c9f17fc3f9feb13"},{"kademlia_liveness_address":"127.0.0.1:13005","kademlia_peer_id":[8,213,222,19,37,27,200,55,153,93,251,67,102,135,69,234,100,7,86,24,117,236,35,185,84,24,216,192,132,244,97,140],"node_id":"node-5","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12005","udp_gossip_address":"127.0.0.1:11005","validator_public_key":"033bf7befb9d7b9e0f8ebf546cd76eee81ec1894912ba6461cc1e85ce9477923ec"}],"genesis-harvesters":[{"kademlia_liveness_address":"127.0.0.1:13002","kademlia_peer_id":[81,120,90,212,28,158,181,19,89,249,118,73,70,102,140,223,197,132,190,161,146,173,165,228,234,94,180,198,252,188,70,237],"node_id":"node-2","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12002","udp_gossip_address":"127.0.0.1:11002","validator_public_key":"022f3b1668765739784259fb1720f2fd367c8f553fab8e1eafa0db3981da5f95ac"},{"kademlia_liveness_address":"127.0.0.1:13003","kademlia_peer_id":[189,65,221,124,105,11,229,100,11,229,165,120,255,142,204,181,200,58,52,85,196,133,202,128,196,128,241,231,29,4,63,143],"node_id":"node-3","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12003","udp_gossip_address":"127.0.0.1:11003","validator_public_key":"02773f2d5b119fd3684665d4edfb724429ac63c139d5fc98891ded955e07ee7603"}],"genesis-miner":{"kademlia_liveness_address":"127.0.0.1:13001","kademlia_peer_id":[142,117,131,195,116,37,160,13,179,42,194,176,55,29,31,234,0,75,72,123,165,226,121,67,114,228,72,224,223,56,186,249],"node_id":"node-1","node_type":"bootstrap","raptorq_gossip_address":"127.0.0.1:12001","udp_gossip_address":"127.0.0.1:11001","validator_public_key":"03d7556dc734cac47037a21f835b50a892f161cd9891c0bf3b38cec40ac43ce128"}}
+{"genesis-farmers":[{"kademlia_liveness_address":"127.0.0.1:13004","kademlia_peer_id":[235,97,10,75,5,33,224,13,151,156,230,90,48,62,79,169,134,37,4,249,112,18,100,70,93,117,65,44,76,205,23,163],"node_id":"node-4","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12004","udp_gossip_address":"127.0.0.1:11004","validator_public_key":"022dea00ded358074aaa154c8c2d1d7e8da3ebe9236996bfbc89c056bb64a836cb"},{"kademlia_liveness_address":"127.0.0.1:13005","kademlia_peer_id":[182,205,113,85,120,184,186,89,54,107,36,133,158,44,225,144,107,198,62,80,217,170,135,89,174,129,210,118,185,23,211,112],"node_id":"node-5","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12005","udp_gossip_address":"127.0.0.1:11005","validator_public_key":"03ceeb5f4ef5bd9da26a20843affc472f7607a4f252d5155b4db642db1d93b6e70"}],"genesis-harvesters":[{"kademlia_liveness_address":"127.0.0.1:13002","kademlia_peer_id":[118,93,242,210,16,231,108,100,29,133,200,70,165,17,99,250,125,208,100,235,196,190,139,150,96,5,45,203,145,119,17,194],"node_id":"node-2","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12002","udp_gossip_address":"127.0.0.1:11002","validator_public_key":"024d29b0fdedc4cea83cd193c6d19f7d6e2ad9867f63e20b38b04c89940dae580f"},{"kademlia_liveness_address":"127.0.0.1:13003","kademlia_peer_id":[71,129,27,71,41,180,173,156,195,236,101,74,54,188,230,116,44,249,15,27,56,79,39,3,45,146,24,7,155,38,75,157],"node_id":"node-3","node_type":"validator","raptorq_gossip_address":"127.0.0.1:12003","udp_gossip_address":"127.0.0.1:11003","validator_public_key":"025145094ff505b7fff78a5a4e7a0e2bc76fb8eb14f5034e3bc9ee42c91138c07d"}],"genesis-miner":{"kademlia_liveness_address":"127.0.0.1:13001","kademlia_peer_id":[19,233,149,239,67,198,170,16,101,31,234,45,204,121,207,203,76,237,129,111,239,20,64,32,71,204,27,193,90,136,164,39],"node_id":"node-1","node_type":"bootstrap","raptorq_gossip_address":"127.0.0.1:12001","udp_gossip_address":"127.0.0.1:11001","validator_public_key":"036711395987e0d57847bab78187a03b2636d6e9d35d36626c1d634a8123ebad68"}}
