+++ crates/consensus/versa_vrf/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "vrrb_vrf"
+name = "versa_vrf"
 version = "0.1.0"
 edition = "2021"
