+++ crates/node/src/runtime/handler_helpers.rs
@@ -7,7 +7,7 @@ use miner::conflict_resolver::Resolver;
 use primitives::{Address, NodeId, NodeType, PublicKey, QuorumId, QuorumKind, Signature};
 use signer::engine::{QuorumData, QuorumMembers as InaugaratedMembers};
 use std::{collections::HashMap, fmt::format};
-use storage::vrrbdb::ApplyBlockResult;
+use storage::versadb::ApplyBlockResult;
 
 use crate::{
     node_runtime::NodeRuntime,
@@ -196,7 +196,6 @@ impl NodeRuntime {
             .ok_or(NodeError::Other(
                 "certificate not appended to convergence block".to_string(),
             ))?;
-        
 
         Ok(block.clone())
     }
