+++ crates/storage/src/lib.rs
@@ -1,2 +1,2 @@
 pub use storage_utils;
-pub use vrrbdb;
+pub use versadb;
