+++ crates/storage/versadb/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "vrrbdb"
+name = "versadb"
 version = "0.1.0"
 edition = "2021"
