+++ crates/storage/versadb/src/lib.rs
@@ -5,19 +5,19 @@ mod state_store;
 pub mod test_utils;
 mod transaction_store;
 pub mod types;
-mod vrrbdb;
-mod vrrbdb_read_handle;
-mod vrrbdb_serialized_values;
+mod versadb;
+mod versadb_read_handle;
+mod versadb_serialized_values;
 
 pub use claim_store::*;
 pub use rocksdb_adapter::*;
 pub use state_store::*;
 pub use transaction_store::*;
 pub use types::*;
-pub use vrrbdb_read_handle::*;
-pub use vrrbdb_serialized_values::*;
+pub use versadb_read_handle::*;
+pub use versadb_serialized_values::*;
 
-pub use crate::vrrbdb::*;
+pub use crate::versadb::*;
 
 #[cfg(test)]
 mod tests {
