+++ crates/storage/versadb/src/result.rs
