+++ crates/versa_rpc/src/rpc/server.rs
@@ -4,7 +4,7 @@ use events::{EventPublisher, DEFAULT_BUFFER};
 use jsonrpsee::server::{ServerBuilder, ServerHandle};
 use mempool::{LeftRightMempool, MempoolReadHandleFactory};
 use primitives::NodeType;
-use storage::vrrbdb::{VrrbDb, VrrbDbConfig, VrrbDbReadHandle};
+use storage::versadb::{VrrbDb, VrrbDbConfig, VrrbDbReadHandle};
 use tokio::sync::mpsc::channel;
 
 use crate::rpc::{api::RpcApiServer, server_impl::RpcServerImpl};
@@ -12,7 +12,7 @@ use crate::rpc::{api::RpcApiServer, server_impl::RpcServerImpl};
 #[derive(Debug, Clone)]
 pub struct JsonRpcServerConfig {
     pub address: SocketAddr,
-    pub vrrbdb_read_handle: VrrbDbReadHandle,
+    pub versadb_read_handle: VrrbDbReadHandle,
     pub mempool_read_handle_factory: MempoolReadHandleFactory,
     pub node_type: NodeType,
     pub events_tx: EventPublisher,
@@ -28,7 +28,7 @@ impl JsonRpcServer {
         let server_impl = RpcServerImpl {
             node_type: config.node_type,
             events_tx: config.events_tx.clone(),
-            vrrbdb_read_handle: config.vrrbdb_read_handle.clone(),
+            versadb_read_handle: config.versadb_read_handle.clone(),
             mempool_read_handle_factory: config.mempool_read_handle_factory.clone(),
         };
 
@@ -45,15 +45,15 @@ impl JsonRpcServer {
 impl Default for JsonRpcServerConfig {
     fn default() -> JsonRpcServerConfig {
         let address = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 9293);
-        let mut vrrbdb_config = VrrbDbConfig::default();
+        let mut versadb_config = VrrbDbConfig::default();
 
         let temp_dir_path = std::env::temp_dir();
         let db_path = temp_dir_path.join(versa_core::helpers::generate_random_string());
 
-        vrrbdb_config.path = db_path;
+        versadb_config.path = db_path;
 
-        let vrrbdb = VrrbDb::new(vrrbdb_config);
-        let vrrbdb_read_handle = vrrbdb.read_handle();
+        let versadb = VrrbDb::new(versadb_config);
+        let versadb_read_handle = versadb.read_handle();
 
         let mempool = LeftRightMempool::default();
         let mempool_read_handle_factory = mempool.factory();
@@ -63,7 +63,7 @@ impl Default for JsonRpcServerConfig {
 
         JsonRpcServerConfig {
             address,
-            vrrbdb_read_handle,
+            versadb_read_handle,
             mempool_read_handle_factory,
             node_type,
             events_tx,
